+++ src/Enchantments.h
@@ -112,7 +112,7 @@ public:
 	// tolua_end
 
 	//* Merge enchantments, check for valid enchantments on item type */
-	int Merge(const cEnchantments & a_Other, int a_ItemType, int a_MergeItemType);
+	int Merge(const cEnchantments & a_Other, short a_ItemType, short a_MergeItemType);
 
 	/** Add enchantment weights from item to the vector */
 	static void AddItemEnchantmentWeights(cWeightedEnchantments & a_Enchantments, short a_ItemType, int a_EnchantmentLevel);
@@ -130,10 +130,10 @@ public:
 	static void CheckEnchantmentConflictsFromVector(cWeightedEnchantments & a_Enchantments, cEnchantments a_FirstEnchantment);
 
 	//** Check if enchantment is valid for item type */
-	static bool IsValidEnchantment(int a_ItemType, int a_EnchantmentID);
+	static bool IsValidEnchantment(short a_ItemType, int a_EnchantmentID);
 
 	//** Check level caps on enchantment */
-	static int GetLevelCap(int a_EnchantmentID, unsigned int a_EnchantmentLevel);
+	static unsigned int GetLevelCap(int a_EnchantmentID, unsigned int a_EnchantmentLevel);
 
 	/** Gets random enchantment from Vector and returns it */
 	static cEnchantments GetRandomEnchantmentFromVector(cWeightedEnchantments & a_Enchantments);
